+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Before every entry put one of these to mark the severity of the change:
 - [Minor] Added sphinx docs.
 - [Minor] Added `with` statement support to EDrivePositioning.
 - [Minor] Added method to `EDriveModbus` to configure the modbus timeout.
+- [Minor] Added method to update I/O data.
+- [Minor] Added methods for reading current position and velocity
 - [Patch] Added graceful closing of ethernetip/modbus session when driver is destructed.
 ### Changed
 - [Minor] Renamed `CMMT` occurences to `EDrive`